split editor.rs to editor/mod.rs and editor.normal_mode, changed CommandLog to KeyStrokeLog
